Begin new effort to encapsulate Runtime and Configuration.
Create scalems.radical.runtime module, and move some definitions from
`scalems/radical/__init__.py`
